Added nowait and split loops up
Added the nowait clause to the pragma omp for command. It currently
hasn’t led to any unusual results, but it could potentially lead to
race conditions. Need to test on server but on local it led to a
speedup.

Split up the x and y limdiff loop into 2 loops to hopefully lead to
some vectorization opportunities by the intel compiler. Will find out
when running it on the cluster